// src/components/ReferanslarSayfasiIsAkisi/types.ts
// This file is auto-generated — do not edit manually.
import type { IkasImage } from "@ikas/bp-storefront";

export interface Props {
  sectionAnchorId?: string;
  workflow4DescriptionHtml?: string;
  workflow4TitleText?: string;
  workflow3ImageUrl?: IkasImage | null;
  workflow3DescriptionHtml?: string;
  workflow3TitleText?: string;
  workflow2ImageUrl?: IkasImage | null;
  workflow2DescriptionHtml?: string;
  workflow2TitleText?: string;
  workflow4ImageUrl?: IkasImage | null;
  workflow1DescriptionHtml?: string;
  workflow1TitleText?: string;
  workflowDescription?: string;
  workflow1ImageUrl?: IkasImage | null;
  workflowTitle?: string;
  workflowKicker?: string;
  workflowDescriptionEn?: string;
  workflow1TitleTextEn?: string;
  workflowKickerEn?: string;
  workflow1DescriptionHtmlEn?: string;
  workflow4DescriptionHtmlEn?: string;
  workflow2TitleTextEn?: string;
  workflow4TitleTextEn?: string;
  workflow2DescriptionHtmlEn?: string;
  workflow3DescriptionHtmlEn?: string;
  workflowTitleEn?: string;
  workflow3TitleTextEn?: string;
  mutedTextColor?: string;
  lineColor?: string;
  paddingBottom?: number;
  paddingTop?: number;
  backgroundColor?: string;
  maxWidth?: number;
  textColor?: string;
  showWorkflowCard2?: boolean;
  showWorkflow?: boolean;
  showWorkflowCard1?: boolean;
  showWorkflowCard3?: boolean;
  showWorkflowCard4?: boolean;
}